add new file for image logic
